# Optimize code using FromMetricsToLabelAdapters