Signed-off-by: Ben Ye <benye@amazon.com>

<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with master
-->

**What this PR does**:

`FromMetricsToLabelAdapters` has some performance penalties as it need to sort all the metrics. That's why we don't want to use it in performance sensitive path. https://github.com/cortexproject/cortex/blob/master/pkg/cortexpb/compat.go#L112

This code stops using it and switches to equivalent code instead.

Benchmark for `SampleStream` `UnmarshalJSON` method before and after this change.

```
name                              old time/op    new time/op    delta
SampleStreamUnmarshalJSONITER-10    7.00ms ± 0%    6.93ms ± 0%   ~     (p=1.000 n=1+1)

name                              old alloc/op   new alloc/op   delta
SampleStreamUnmarshalJSONITER-10    5.46MB ± 0%    4.17MB ± 0%   ~     (p=1.000 n=1+1)

name                              old allocs/op  new allocs/op  delta
SampleStreamUnmarshalJSONITER-10     90.3k ± 0%     60.3k ± 0%   ~     (p=1.000 n=1+1)
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`